+++ tests/models/decoder_only/vision_language/test_models.py
@@ -95,7 +95,7 @@ VLM_TEST_SETTINGS = {
             limit_mm_per_prompt={"image": 4},
         )],
         # TODO: remove this after CPU float16 support ready
-        dtype="float" if current_platform.is_cpu() else "half",
+        dtype="bfloat16" if current_platform.is_cpu() else "half",
         marks=[pytest.mark.core_model],
     ),
     "paligemma": VLMTestInfo(
@@ -134,7 +134,7 @@ VLM_TEST_SETTINGS = {
         auto_cls=AutoModelForVision2Seq,
         vllm_output_post_proc=model_utils.qwen2_vllm_to_hf_output,
         # TODO: remove this after CPU float16 support ready
-        dtype="float" if current_platform.is_cpu() else "half",
+        dtype="bfloat16" if current_platform.is_cpu() else "half",
         marks=[pytest.mark.core_model],
         image_size_factors=[(), (0.25,), (0.25, 0.25, 0.25), (0.25, 0.2, 0.15)],
     ),
